Move ModeratorToolsModal to separate file
